fix poetry entrypoint definition
